stderr >> buffer

+++ gofiles.go
@@ -1,14 +1,26 @@
 package main
 
 import (
+	"bytes"
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"os/exec"
 	"time"
 
 	c "github.com/Joker/ioterm"
+	// "github.com/k0kubun/pp"
 )
+
+type panicWriter struct {
+	w io.Writer
+}
+
+func (pw panicWriter) Write(p []byte) (int, error) {
+	pw.w.Write([]byte("\033[31m"))
+	return pw.w.Write(p)
+}
 
 func (dot *project) make() {
 	for {
@@ -31,7 +43,7 @@ func (dot *project) make() {
 				continue
 			}
 
-			dot.start()
+			go dot.start()
 		} else {
 			break
 		}
@@ -43,12 +55,31 @@ func (dot *project) start() {
 
 	dot.cmd = exec.Command("./" + dot.name)
 	dot.cmd.Stdout = os.Stdout
-	dot.cmd.Stderr = os.Stderr
-	dot.cmd.Start()
+	// dot.cmd.Stderr = panicWriter{os.Stderr}
+	stderr, err := dot.cmd.StderrPipe()
+	if err != nil {
+		log.Println("Error cmd.StderrPipe() - ", err)
+	}
+
+	err = dot.cmd.Start()
+	if err != nil {
+		log.Println("Error cmd.Start() - ", err)
+	}
+
+	sebuf := new(bytes.Buffer)
+	go io.Copy(sebuf, stderr)
+
+	err = dot.cmd.Wait()
+	if err != nil {
+		if !err.(*exec.ExitError).Success() {
+			fmt.Println(c.Blue_l, sebuf, c.Reset)
+			c.Errorf("%s pid:%d -- %v", dot.name, err.(*exec.ExitError).Pid(), err)
+		}
+	}
 }
 
 func (dot *project) stop() {
-	if dot.cmd != nil && dot.cmd.Process != nil {
+	if dot.cmd != nil && dot.cmd.Process != nil && !dot.cmd.ProcessState.Exited() {
 		fmt.Println(c.Red_h, "--  stop --", c.Reset)
 		err := dot.cmd.Process.Kill()
 		if err != nil {
